Ignoring HTML debris
Let's address the major drawback: HTML junk in our content.

Adding a `plain_text_body` column to our `action_text_rich_texts` table
---

As a first measure, we'll add a secondary `plain_text_body` column that
mirrors the existing `body` column so that our `ActionText::RichText`
models can write to and search from it. We'll use Rails' `migration`
generator to create the migration:

```ruby
class AddPlainTextBodyToActionTextRichTexts < ActiveRecord::Migration[7.0]
  def change
    add_column :action_text_rich_texts, :plain_text_body, :text
  end
end
```

Writing to `action_text_rich_texts.plain_text_body`
---

To write to the new column, we'll declare a [before_save][] callback so
that creating or updating an `Article` instance transforms by invoking
[ActionText::RichText#to_plain_text][]. The call to `to_plain_text` will
remove all HTML syntax, leaving behind only the text content of the
`body` to be written to `plain_text_body`:

```diff
--- a/app/models/article.rb
+++ b/app/models/article.rb
 class Article < ApplicationRecord
   has_rich_text :content
+
+  before_save { content.plain_text_body = content.body.to_plain_text }

   scope :with_content_containing, ->(query) { joins(:rich_text_content).merge(ActionText::RichText.where <<~SQL, "%" + query + "%") }
     body ILIKE ?
   SQL
 end
```

Querying `action_text_rich_texts.plain_text_body`
---

Now that we have a column free of HTML entities, we can point our
existing query to it in support querying of values like `<div>The text
of this article matches has the <strong>needle</strong> in the
haystack</div>`.

While this is a major improvement in what's _possible_ to query, there
are still some quirks and drawbacks. For instance, querying with `ILIKE`
does not account for "stop words" like "in", "of", or "the". This means
that while a query like `"needle in the haystack"` will match `"The text
of this article matches has the needle in the haystack"`, `"needle
haystack"` will not.

```diff
--- a/app/models/article.rb
+++ b/app/models/article.rb
 class Article < ApplicationRecord
   has_rich_text :content

   before_save { content.plain_text_body = content.body.to_plain_text }

   scope :with_content_containing, ->(query) { joins(:rich_text_content).merge(ActionText::RichText.where <<~SQL, "%" + query + "%") }
-    body ILIKE ?
+    plain_text_body ILIKE ?
   SQL
 end
```

[before_save]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/Callbacks.html
[ActionText::RichText#to_plain_text]: https://edgeapi.rubyonrails.org/classes/ActionText/RichText.html#method-i-to_plain_text